--- src/main/java/de/ralleytn/simple/json/JSONArray.java
@@ -1,999 +1,0 @@
-/*
- *                                  Apache License
- *                            Version 2.0, January 2004
- *                         http://www.apache.org/licenses/
- * 
- *    TERMS AND CONDITIONS FOR USE, REPRODUCTION, AND DISTRIBUTION
- * 
- *    1. Definitions.
- * 
- *       "License" shall mean the terms and conditions for use, reproduction,
- *       and distribution as defined by Sections 1 through 9 of this document.
- * 
- *       "Licensor" shall mean the copyright owner or entity authorized by
- *       the copyright owner that is granting the License.
- * 
- *       "Legal Entity" shall mean the union of the acting entity and all
- *       other entities that control, are controlled by, or are under common
- *       control with that entity. For the purposes of this definition,
- *       "control" means (i) the power, direct or indirect, to cause the
- *       direction or management of such entity, whether by contract or
- *       otherwise, or (ii) ownership of fifty percent (50%) or more of the
- *       outstanding shares, or (iii) beneficial ownership of such entity.
- * 
- *       "You" (or "Your") shall mean an individual or Legal Entity
- *       exercising permissions granted by this License.
- * 
- *       "Source" form shall mean the preferred form for making modifications,
- *       including but not limited to software source code, documentation
- *       source, and configuration files.
- * 
- *       "Object" form shall mean any form resulting from mechanical
- *       transformation or translation of a Source form, including but
- *       not limited to compiled object code, generated documentation,
- *       and conversions to other media types.
- * 
- *       "Work" shall mean the work of authorship, whether in Source or
- *       Object form, made available under the License, as indicated by a
- *       copyright notice that is included in or attached to the work
- *       (an example is provided in the Appendix below).
- * 
- *       "Derivative Works" shall mean any work, whether in Source or Object
- *       form, that is based on (or derived from) the Work and for which the
- *       editorial revisions, annotations, elaborations, or other modifications
- *       represent, as a whole, an original work of authorship. For the purposes
- *       of this License, Derivative Works shall not include works that remain
- *       separable from, or merely link (or bind by name) to the interfaces of,
- *       the Work and Derivative Works thereof.
- * 
- *       "Contribution" shall mean any work of authorship, including
- *       the original version of the Work and any modifications or additions
- *       to that Work or Derivative Works thereof, that is intentionally
- *       submitted to Licensor for inclusion in the Work by the copyright owner
- *       or by an individual or Legal Entity authorized to submit on behalf of
- *       the copyright owner. For the purposes of this definition, "submitted"
- *       means any form of electronic, verbal, or written communication sent
- *       to the Licensor or its representatives, including but not limited to
- *       communication on electronic mailing lists, source code control systems,
- *       and issue tracking systems that are managed by, or on behalf of, the
- *       Licensor for the purpose of discussing and improving the Work, but
- *       excluding communication that is conspicuously marked or otherwise
- *       designated in writing by the copyright owner as "Not a Contribution."
- * 
- *       "Contributor" shall mean Licensor and any individual or Legal Entity
- *       on behalf of whom a Contribution has been received by Licensor and
- *       subsequently incorporated within the Work.
- * 
- *    2. Grant of Copyright License. Subject to the terms and conditions of
- *       this License, each Contributor hereby grants to You a perpetual,
- *       worldwide, non-exclusive, no-charge, royalty-free, irrevocable
- *       copyright license to reproduce, prepare Derivative Works of,
- *       publicly display, publicly perform, sublicense, and distribute the
- *       Work and such Derivative Works in Source or Object form.
- * 
- *    3. Grant of Patent License. Subject to the terms and conditions of
- *       this License, each Contributor hereby grants to You a perpetual,
- *       worldwide, non-exclusive, no-charge, royalty-free, irrevocable
- *       (except as stated in this section) patent license to make, have made,
- *       use, offer to sell, sell, import, and otherwise transfer the Work,
- *       where such license applies only to those patent claims licensable
- *       by such Contributor that are necessarily infringed by their
- *       Contribution(s) alone or by combination of their Contribution(s)
- *       with the Work to which such Contribution(s) was submitted. If You
- *       institute patent litigation against any entity (including a
- *       cross-claim or counterclaim in a lawsuit) alleging that the Work
- *       or a Contribution incorporated within the Work constitutes direct
- *       or contributory patent infringement, then any patent licenses
- *       granted to You under this License for that Work shall terminate
- *       as of the date such litigation is filed.
- * 
- *    4. Redistribution. You may reproduce and distribute copies of the
- *       Work or Derivative Works thereof in any medium, with or without
- *       modifications, and in Source or Object form, provided that You
- *       meet the following conditions:
- * 
- *       (a) You must give any other recipients of the Work or
- *           Derivative Works a copy of this License; and
- * 
- *       (b) You must cause any modified files to carry prominent notices
- *           stating that You changed the files; and
- * 
- *       (c) You must retain, in the Source form of any Derivative Works
- *           that You distribute, all copyright, patent, trademark, and
- *           attribution notices from the Source form of the Work,
- *           excluding those notices that do not pertain to any part of
- *           the Derivative Works; and
- * 
- *       (d) If the Work includes a "NOTICE" text file as part of its
- *           distribution, then any Derivative Works that You distribute must
- *           include a readable copy of the attribution notices contained
- *           within such NOTICE file, excluding those notices that do not
- *           pertain to any part of the Derivative Works, in at least one
- *           of the following places: within a NOTICE text file distributed
- *           as part of the Derivative Works; within the Source form or
- *           documentation, if provided along with the Derivative Works; or,
- *           within a display generated by the Derivative Works, if and
- *           wherever such third-party notices normally appear. The contents
- *           of the NOTICE file are for informational purposes only and
- *           do not modify the License. You may add Your own attribution
- *           notices within Derivative Works that You distribute, alongside
- *           or as an addendum to the NOTICE text from the Work, provided
- *           that such additional attribution notices cannot be construed
- *           as modifying the License.
- * 
- *       You may add Your own copyright statement to Your modifications and
- *       may provide additional or different license terms and conditions
- *       for use, reproduction, or distribution of Your modifications, or
- *       for any such Derivative Works as a whole, provided Your use,
- *       reproduction, and distribution of the Work otherwise complies with
- *       the conditions stated in this License.
- * 
- *    5. Submission of Contributions. Unless You explicitly state otherwise,
- *       any Contribution intentionally submitted for inclusion in the Work
- *       by You to the Licensor shall be under the terms and conditions of
- *       this License, without any additional terms or conditions.
- *       Notwithstanding the above, nothing herein shall supersede or modify
- *       the terms of any separate license agreement you may have executed
- *       with Licensor regarding such Contributions.
- * 
- *    6. Trademarks. This License does not grant permission to use the trade
- *       names, trademarks, service marks, or product names of the Licensor,
- *       except as required for reasonable and customary use in describing the
- *       origin of the Work and reproducing the content of the NOTICE file.
- * 
- *    7. Disclaimer of Warranty. Unless required by applicable law or
- *       agreed to in writing, Licensor provides the Work (and each
- *       Contributor provides its Contributions) on an "AS IS" BASIS,
- *       WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
- *       implied, including, without limitation, any warranties or conditions
- *       of TITLE, NON-INFRINGEMENT, MERCHANTABILITY, or FITNESS FOR A
- *       PARTICULAR PURPOSE. You are solely responsible for determining the
- *       appropriateness of using or redistributing the Work and assume any
- *       risks associated with Your exercise of permissions under this License.
- * 
- *    8. Limitation of Liability. In no event and under no legal theory,
- *       whether in tort (including negligence), contract, or otherwise,
- *       unless required by applicable law (such as deliberate and grossly
- *       negligent acts) or agreed to in writing, shall any Contributor be
- *       liable to You for damages, including any direct, indirect, special,
- *       incidental, or consequential damages of any character arising as a
- *       result of this License or out of the use or inability to use the
- *       Work (including but not limited to damages for loss of goodwill,
- *       work stoppage, computer failure or malfunction, or any and all
- *       other commercial damages or losses), even if such Contributor
- *       has been advised of the possibility of such damages.
- * 
- *    9. Accepting Warranty or Additional Liability. While redistributing
- *       the Work or Derivative Works thereof, You may choose to offer,
- *       and charge a fee for, acceptance of support, warranty, indemnity,
- *       or other liability obligations and/or rights consistent with this
- *       License. However, in accepting such obligations, You may act only
- *       on Your own behalf and on Your sole responsibility, not on behalf
- *       of any other Contributor, and only if You agree to indemnify,
- *       defend, and hold each Contributor harmless for any liability
- *       incurred by, or claims asserted against, such Contributor by reason
- *       of your accepting any such warranty or additional liability.
- * 
- *    END OF TERMS AND CONDITIONS
- * 
- *    APPENDIX: How to apply the Apache License to your work.
- * 
- *       To apply the Apache License to your work, attach the following
- *       boilerplate notice, with the fields enclosed by brackets "{}"
- *       replaced with your own identifying information. (Don't include
- *       the brackets!)  The text should be enclosed in the appropriate
- *       comment syntax for the file format. We also recommend that a
- *       file or class name and description of purpose be included on the
- *       same "printed page" as the copyright notice for easier
- *       identification within third-party archives.
- * 
- *    Copyright {yyyy} {name of copyright owner}
- * 
- *    Licensed under the Apache License, Version 2.0 (the "License");
- *    you may not use this file except in compliance with the License.
- *    You may obtain a copy of the License at
- * 
- *        http://www.apache.org/licenses/LICENSE-2.0
- * 
- *    Unless required by applicable law or agreed to in writing, software
- *    distributed under the License is distributed on an "AS IS" BASIS,
- *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *    See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package de.ralleytn.simple.json;
-
-import java.io.IOException;
-import java.io.Reader;
-import java.io.StringWriter;
-import java.io.Writer;
-import java.lang.reflect.Array;
-import java.text.DateFormat;
-import java.text.ParseException;
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.Date;
-import java.util.Map;
-
-import de.ralleytn.simple.json.internal.Util;
-
-/**
- * Represents a JSON array.
- * @author FangYidong(fangyidong@yahoo.com.cn)
- * @author Ralph Niemitz/RalleYTN(ralph.niemitz@gmx.de)
- * @version 2.0.0
- * @since 1.0.0
- */
-public class JSONArray extends ArrayList<Object> {
-	
-	// ==== 11.03.2018 | Ralph Niemitz/RalleYTN(ralph.niemitz@gmx.de)
-	// -	Removed the interfaces JSONAware and JSONStreamAware
-	// -	Moved all static "writeJSONString" methods to the Util class
-	// -	Removed all "toJSONString" methods
-	// -	Renamed the "writeJSONString" method to simply "write"
-	// -	Updated documentation
-	// ====
-	
-	private static final long serialVersionUID = 3957988303675231981L;
-	
-	/**
-	 * Constructs an empty {@linkplain JSONArray}.
-	 * @since 1.0.0
-	 */
-	public JSONArray() {}
-	
-	/**
-	 * Constructs a {@linkplain JSONArray} with the elements of the given {@linkplain Collection}.
-	 * @param collection the {@linkplain Collection}
-	 * @since 1.0.0
-	 */
-	public JSONArray(Collection<?> collection){
-		
-		super(collection);
-	}
-	
-	/**
-	 * Constructs a {@linkplain JSONArray} with the elements of the given array.
-	 * @param array the array
-	 * @param <T> the array type
-	 * @since 1.0.0
-	 */
-	public <T>JSONArray(T[] array) {
-		
-		for(T element : array) {
-			
-			this.add(element);
-		}
-	}
-	
-	/**
-	 * Constructs a {@linkplain JSONArray} with the elements of the given array.
-	 * @param array the array
-	 * @since 1.0.0
-	 */
-	public JSONArray(byte[] array) {
-		
-		for(byte element : array) {
-			
-			this.add(element);
-		}
-	}
-	
-	/**
-	 * Constructs a {@linkplain JSONArray} with the elements of the given array.
-	 * @param array the array
-	 * @since 1.0.0
-	 */
-	public JSONArray(boolean[] array) {
-		
-		for(boolean element : array) {
-			
-			this.add(element);
-		}
-	}
-	
-	/**
-	 * Constructs a {@linkplain JSONArray} with the elements of the given array.
-	 * @param array the array
-	 * @since 1.0.0
-	 */
-	public JSONArray(char[] array) {
-		
-		for(char element : array) {
-			
-			this.add(element);
-		}
-	}
-	
-	/**
-	 * Constructs a {@linkplain JSONArray} with the elements of the given array.
-	 * @param array the array
-	 * @since 1.0.0
-	 */
-	public JSONArray(short[] array) {
-		
-		for(short element : array) {
-			
-			this.add(element);
-		}
-	}
-	
-	/**
-	 * Constructs a {@linkplain JSONArray} with the elements of the given array.
-	 * @param array the array
-	 * @since 1.0.0
-	 */
-	public JSONArray(int[] array) {
-		
-		for(int element : array) {
-			
-			this.add(element);
-		}
-	}
-	
-	/**
-	 * Constructs a {@linkplain JSONArray} with the elements of the given array.
-	 * @param array the array
-	 * @since 1.0.0
-	 */
-	public JSONArray(long[] array) {
-		
-		for(long element : array) {
-			
-			this.add(element);
-		}
-	}
-	
-	/**
-	 * Constructs a {@linkplain JSONArray} with the elements of the given array.
-	 * @param array the array
-	 * @since 1.0.0
-	 */
-	public JSONArray(float[] array) {
-		
-		for(float element : array) {
-			
-			this.add(element);
-		}
-	}
-	
-	/**
-	 * Constructs a {@linkplain JSONArray} with the elements of the given array.
-	 * @param array the array
-	 * @since 1.0.0
-	 */
-	public JSONArray(double[] array) {
-		
-		for(double element : array) {
-			
-			this.add(element);
-		}
-	}
-	
-	/**
-	 * Constructs a {@linkplain JSONArray} with the elements of the given array.
-	 * This constructor is a little slower than the others because it uses the Reflection API.
-	 * @param array the array
-	 * @since 1.0.0
-	 */
-	public JSONArray(Object array) {
-		
-		// ==== 11.03.2018 | Ralph Niemitz/RalleYTN(ralph.niemitz@gmx.de)
-		// -	Will no longer throw an exception if an object that isn't an array is given.
-		//		An empty array will be constructed instead.
-		// ====
-		
-		if(array != null && array.getClass().isArray()) {
-			
-			int length = Array.getLength(array);
-			
-			for(int index = 0; index < length; index++) {
-				
-				this.add(Array.get(array, index));
-			}
-		}
-	}
-	
-	/**
-	 * Constructs a {@linkplain JSONArray} from a JSON string.
-	 * @param json the JSON string
-	 * @throws JSONParseException if the JSON is invalid
-	 * @since 1.0.0
-	 */
-	public JSONArray(String json) throws JSONParseException {
-		
-		super((JSONArray)new JSONParser().parse(json));
-	}
-	
-	/**
-	 * Constructs a {@linkplain JSONArray} from JSON data read from a {@linkplain Reader}.
-	 * @param reader the {@linkplain Reader}
-	 * @throws JSONParseException if the JSON is invalid
-	 * @throws IOException if an I/O error occurred
-	 * @since 1.0.0
-	 */
-	public JSONArray(Reader reader) throws JSONParseException, IOException {
-		
-		super((JSONArray)new JSONParser().parse(reader));
-	}
-	
-	/**
-	 * Writes this {@linkplain JSONArray} as a JSON string on the given {@linkplain Writer}.
-	 * @param writer the {@linkplain Writer}
-	 * @throws IOException if an I/O error occurred
-	 * @since 1.0.0
-	 */
-	public void write(Writer writer) throws IOException {
-		
-		Util.write(this, writer);
-	}
-
-	/**
-	 * Converts this {@linkplain JSONArray} to a JSON string.
-	 * @return this {@linkplain JSONArray} as a JSON string
-	 * @since 1.0.0
-	 */
-	@Override
-	public String toString() {
-		
-		try(StringWriter writer = new StringWriter()) {
-			
-			Util.write(this, writer);
-			return writer.toString();
-			
-		} catch(IOException exception) {
-
-			// WILL NEVER HAPPEN!
-			// DO NOTHING!
-		}
-		
-		return null;
-	}
-	
-	@Override
-	public boolean equals(Object object) {
-		
-		if(object != null) {
-			
-			if(object instanceof Collection) {
-				
-				Collection<?> collection = (Collection<?>)object;
-				
-				if(collection.size() == this.size()) {
-					
-					int index = 0;
-					
-					for(Object element : collection) {
-						
-						if(!((element == null && this.get(index) == null) || this.get(index).equals(element))) {
-							
-							return false;
-						}
-						
-						index++;
-					}
-					
-					return true;
-				}
-				
-			} else if(object.getClass().isArray()) {
-				
-				int length = Array.getLength(object);
-				
-				if(length == this.size()) {
-					
-					for(int index = 0; index < length; index++) {
-						
-						Object element = Array.get(object, index);
-
-						if(!((element == null && this.get(index) == null) || element.equals(this.get(index)))) {
-							
-							return false;
-						}
-					}
-					
-					return true;
-				}
-			}
-		}
-		
-		return false;
-	}
-	
-	/**
-	 * If the value is a {@linkplain JSONObject} already, it will be casted and returned.
-	 * If the value is a {@linkplain Map}, it will be wrapped in a {@linkplain JSONObject}. The wrapped {@linkplain JSONObject} will be returned.
-	 * In any other case this method returns {@code null}.
-	 * @param index index of the value
-	 * @return a {@code JSONObject} or {@code null}
-	 * @since 1.0.0
-	 */
-	public JSONObject getObject(int index) {
-		
-		return Util.getObject(this.get(index));
-	}
-	
-	/**
-	 * If the value already is a {@linkplain JSONArray}, it will be casted and returned.
-	 * If the value is an array or {@linkplain Collection}, it will be wrapped in a {@linkplain JSONArray}.
-	 * The {@linkplain JSONArray} will be returned.
-	 * In any other case this method returns {@code null}.
-	 * @param index index of the value
-	 * @return a {@linkplain JSONArray} or {@code null}
-	 * @since 1.0.0
-	 */
-	public JSONArray getArray(int index) {
-		
-		return Util.getArray(this.get(index));
-	}
-	
-	/**
-	 * If the value is already a {@linkplain Boolean}, it will be casted and returned.
-	 * If the value is a {@linkplain String}, it will be parsed. The result is returned.
-	 * If the value is a {@linkplain Number}, this method will return {@code true} in case its {@code long} value is {@code 1}.
-	 * It will return {@code false} if it is a number but the {@code long} value is <u>not</u> {@code 1}.
-	 * In any other case this method returns {@code null}.
-	 * @param index index of the value
-	 * @return a {@linkplain Boolean} or {@code null}
-	 * @since 1.0.0
-	 */
-	public Boolean getBoolean(int index) {
-		
-		return Util.getBoolean(this.get(index));
-	}
-	
-	/**
-	 * If the value is a {@linkplain Number}, its {@code byte} value is returned.
-	 * If the value is a {@linkplain String}, it will be parsed. The result is returned.
-	 * If the value is a {@linkplain Boolean}, this method returns {@code 1} in case the value is {@code true} otherwise {@code 0}.
-	 * In any other case this method returns {@code null}.
-	 * @param index index of the value
-	 * @return a {@linkplain Byte} or {@code null}
-	 * @since 1.0.0
-	 */
-	public Byte getByte(int index) {
-		
-		return Util.getByte(this.get(index));
-	}
-	
-	/**
-	 * If the value is a {@linkplain Number}, its {@code short} value is returned.
-	 * If the value is a {@linkplain String}, it will be parsed. The result is returned.
-	 * If the value is a {@linkplain Boolean}, this method returns {@code 1} in case the value is {@code true} otherwise {@code 0}.
-	 * In any other case this method returns {@code null}.
-	 * @param index index of the value
-	 * @return a {@linkplain Short} or {@code null}
-	 * @since 1.0.0
-	 */
-	public Short getShort(int index) {
-		
-		return Util.getShort(this.get(index));
-	}
-	
-	/**
-	 * If the value is a {@linkplain Number}, its {@code int} value is returned.
-	 * If the value is a {@linkplain String}, it will be parsed. The result is returned.
-	 * If the value is a {@linkplain Boolean}, this method returns {@code 1} in case the value is {@code true} otherwise {@code 0}.
-	 * In any other case this method returns {@code null}.
-	 * @param index index of the value
-	 * @return a {@linkplain Integer} or {@code null}
-	 * @since 1.0.0
-	 */
-	public Integer getInteger(int index) {
-		
-		return Util.getInteger(this.get(index));
-	}
-	
-	/**
-	 * If the value is a {@linkplain Number}, its {@code long} value is returned.
-	 * If the value is a {@linkplain String}, it will be parsed. The result is returned.
-	 * If the value is a {@linkplain Boolean}, this method returns {@code 1} in case the value is {@code true} otherwise {@code 0}.
-	 * In any other case this method returns {@code null}.
-	 * @param index index of the value
-	 * @return a {@linkplain Long} or {@code null}
-	 * @since 1.0.0
-	 */
-	public Long getLong(int index) {
-		
-		return Util.getLong(this.get(index));
-	}
-	
-	/**
-	 * If the value is a {@linkplain Number}, its {@code float} value is returned.
-	 * If the value is a {@linkplain String}, it will be parsed. The result is returned.
-	 * If the value is a {@linkplain Boolean}, this method returns {@code 1} in case the value is {@code true} otherwise {@code 0}.
-	 * In any other case this method returns {@code null}.
-	 * @param index index of the value
-	 * @return a {@linkplain Float} or {@code null}
-	 * @since 1.0.0
-	 */
-	public Float getFloat(int index) {
-		
-		return Util.getFloat(this.get(index));
-	}
-	
-	/**
-	 * If the value is a {@linkplain Number}, its {@code double} value is returned.
-	 * If the value is a {@linkplain String}, it will be parsed. The result is returned.
-	 * If the value is a {@linkplain Boolean}, this method returns {@code 1} in case the value is {@code true} otherwise {@code 0}.
-	 * In any other case this method returns {@code null}.
-	 * @param index index of the value
-	 * @return a {@linkplain Double} or {@code null}
-	 * @since 1.0.0
-	 */
-	public Double getDouble(int index) {
-		
-		return Util.getDouble(this.get(index));
-	}
-	
-	/**
-	 * Returns the same as the value's {@link Object#toString()} method.
-	 * If the actual value is {@code null}, this method will return {@code null}.
-	 * @param index index of the value
-	 * @return a {@linkplain String} or {@code null}
-	 * @since 1.0.0
-	 */
-	public String getString(int index) {
-		
-		return Util.getString(this.get(index));
-	}
-
-	/**
-	 * If the value already is a {@linkplain Date}, it will be casted and returned.
-	 * Otherwise the result of the value's {@link Object#toString()} will be parsed by the given
-	 * {@linkplain DateFormat}. The result is returned.
-	 * If the actual value is {@code null}, this method will return {@code null}.
-	 * @param index index of the value
-	 * @param format the {@linkplain DateFormat} used to parse the date
-	 * @return a {@linkplain Date} or {@code null}
-	 * @throws ParseException if the date could not be parsed
-	 * @since 1.0.0
-	 */
-	public Date getDate(int index, DateFormat format) throws ParseException {
-		
-		return Util.getDate(this.get(index), format);
-	}
-	
-	/**
-	 * If the {@linkplain String} representation of the value equals the name of the enum constant
-	 * in the given enum type, it will return the enum constant.
-	 * In any other case this method returns {@code null}.
-	 * @param index index of the value
-	 * @param type the enum type
-	 * @param <T> the return type
-	 * @return an {@linkplain Enum} or {@code null}
-	 * @since 1.0.0
-	 */
-	public <T extends Enum<T>>T getEnum(int index, Class<T> type) {
-		
-		return Util.getEnum(this.get(index), type);
-	}
-
-	/**
-	 * Converts this {@linkplain JSONArray} into an array of {@linkplain Boolean}s using the {@link #getBoolean(int)} method.
-	 * @return an array {@linkplain Boolean}s
-	 * @since 1.0.0
-	 */
-	public Boolean[] toBooleanArray() {
-		
-		Boolean[] array = new Boolean[this.size()];
-		
-		for(int index = 0; index < array.length; index++) {
-			
-			array[index] = this.getBoolean(index);
-		}
-		
-		return array;
-	}
-
-	/**
-	 * Converts this {@linkplain JSONArray} into an array of {@linkplain Byte}s using the {@link #getByte(int)} method.
-	 * @return an array of {@linkplain Byte}s
-	 * @since 1.0.0
-	 */
-	public Byte[] toByteArray() {
-		
-		Byte[] array = new Byte[this.size()];
-		
-		for(int index = 0; index < array.length; index++) {
-			
-			array[index] = this.getByte(index);
-		}
-		
-		return array;
-	}
-
-	/**
-	 * Converts this {@linkplain JSONArray} into an array of {@linkplain Short}s using the {@link #getShort(int)} method.
-	 * @return an array of {@linkplain Short}s
-	 * @since 1.0.0
-	 */
-	public Short[] toShortArray() {
-		
-		Short[] array = new Short[this.size()];
-		
-		for(int index = 0; index < array.length; index++) {
-			
-			array[index] = this.getShort(index);
-		}
-		
-		return array;
-	}
-
-	/**
-	 * Converts this {@linkplain JSONArray} into an array of {@linkplain Integer}s using the {@link #getInteger(int)} method.
-	 * @return an array of {@linkplain Integer}s
-	 * @since 1.0.0
-	 */
-	public Integer[] toIntegerArray() {
-		
-		Integer[] array = new Integer[this.size()];
-		
-		for(int index = 0; index < array.length; index++) {
-			
-			array[index] = this.getInteger(index);
-		}
-		
-		return array;
-	}
-
-	/**
-	 * Converts this {@linkplain JSONArray} into an array of {@linkplain Long}s using the {@link #getLong(int)} method.
-	 * @return an array of {@linkplain Long}s
-	 * @since 1.0.0
-	 */
-	public Long[] toLongArray() {
-		
-		Long[] array = new Long[this.size()];
-		
-		for(int index = 0; index < array.length; index++) {
-			
-			array[index] = this.getLong(index);
-		}
-		
-		return array;
-	}
-
-	/**
-	 * Converts this {@linkplain JSONArray} into an array of {@linkplain Float}s using the {@link #getFloat(int)} method.
-	 * @return an array of {@linkplain Float}s
-	 * @since 1.0.0
-	 */
-	public Float[] toFloatArray() {
-		
-		Float[] array = new Float[this.size()];
-		
-		for(int index = 0; index < array.length; index++) {
-			
-			array[index] = this.getFloat(index);
-		}
-		
-		return array;
-	}
-	
-	/**
-	 * Converts this {@linkplain JSONArray} into an array of {@linkplain Double}s using the {@link #getDouble(int)} method.
-	 * @return an array of {@linkplain Double}s
-	 * @since 1.0.0
-	 */
-	public Double[] toDoubleArray() {
-		
-		Double[] array = new Double[this.size()];
-		
-		for(int index = 0; index < array.length; index++) {
-			
-			array[index] = this.getDouble(index);
-		}
-		
-		return array;
-	}
-	
-	/**
-	 * Converts this {@linkplain JSONArray} into an array of {@linkplain String}s using the {@link #getString(int)} method.
-	 * @return an array of {@linkplain String}s
-	 * @since 1.0.0
-	 */
-	public String[] toStringArray() {
-		
-		String[] array = new String[this.size()];
-		
-		for(int index = 0; index < array.length; index++) {
-			
-			array[index] = this.getString(index);
-		}
-		
-		return array;
-	}
-	
-	/**
-	 * Converts this {@linkplain JSONArray} into an array of {@linkplain JSONObject}s using the {@link #getObject(int)} method.
-	 * @return an array of {@linkplain JSONObject}s
-	 * @since 1.0.0
-	 */
-	public JSONObject[] toObjectArray() {
-		
-		JSONObject[] array = new JSONObject[this.size()];
-		
-		for(int index = 0; index < array.length; index++) {
-			
-			array[index] = this.getObject(index);
-		}
-		
-		return array;
-	}
-	
-	/**
-	 * Converts this {@linkplain JSONArray} into an array of {@linkplain JSONArray}s using the {@link #getArray(int)} method.
-	 * @return an array of {@linkplain JSONArray}s.
-	 * @since 1.0.0
-	 */
-	public JSONArray[] toArrayArray() {
-		
-		JSONArray[] array = new JSONArray[this.size()];
-		
-		for(int index = 0; index < array.length; index++) {
-			
-			array[index] = this.getArray(index);
-		}
-		
-		return array;
-	}
-	
-	/**
-	 * Converts this {@linkplain JSONArray} to a primitive {@code byte} array.
-	 * @return a primitive {@code byte} array
-	 * @since 2.0.0
-	 */
-	public byte[] toPrimitiveByteArray() {
-		
-		byte[] array = new byte[this.size()];
-		
-		for(int index = 0; index < array.length; index++) {
-			
-			array[index] = this.getByte(index);
-		}
-		
-		return array;
-	}
-	
-	/**
-	 * Converts this {@linkplain JSONArray} to a primitive {@code short} array.
-	 * @return a primitive {@code short} array
-	 * @since 2.0.0
-	 */
-	public short[] toPrimitiveShortArray() {
-		
-		short[] array = new short[this.size()];
-		
-		for(int index = 0; index < array.length; index++) {
-			
-			array[index] = this.getShort(index);
-		}
-		
-		return array;
-	}
-	
-	/**
-	 * Converts this {@linkplain JSONArray} to a primitive {@code int} array.
-	 * @return a primitive {@code int} array
-	 * @since 2.0.0
-	 */
-	public int[] toPrimitiveIntArray() {
-		
-		int[] array = new int[this.size()];
-		
-		for(int index = 0; index < array.length; index++) {
-			
-			array[index] = this.getInteger(index);
-		}
-		
-		return array;
-	}
-	
-	/**
-	 * Converts this {@linkplain JSONArray} to a primitive {@code long} array.
-	 * @return a primitive {@code long} array
-	 * @since 2.0.0
-	 */
-	public long[] toPrimitiveLongArray() {
-		
-		long[] array = new long[this.size()];
-		
-		for(int index = 0; index < array.length; index++) {
-			
-			array[index] = this.getLong(index);
-		}
-		
-		return array;
-	}
-	
-	/**
-	 * Converts this {@linkplain JSONArray} to a primitive {@code float} array.
-	 * @return a primitive {@code float} array
-	 * @since 2.0.0
-	 */
-	public float[] toPrimitiveFloatArray() {
-		
-		float[] array = new float[this.size()];
-		
-		for(int index = 0; index < array.length; index++) {
-			
-			array[index] = this.getFloat(index);
-		}
-		
-		return array;
-	}
-	
-	/**
-	 * Converts this {@linkplain JSONArray} to a primitive {@code double} array.
-	 * @return a primitive {@code double} array
-	 * @since 2.0.0
-	 */
-	public double[] toPrimitiveDoubleArray() {
-		
-		double[] array = new double[this.size()];
-		
-		for(int index = 0; index < array.length; index++) {
-			
-			array[index] = this.getDouble(index);
-		}
-		
-		return array;
-	}
-	
-	/**
-	 * Converts this {@linkplain JSONArray} to a primitive {@code boolean} array.
-	 * @return a primitive {@code boolean} array
-	 * @since 2.0.0
-	 */
-	public boolean[] toPrimitiveBooleanArray() {
-		
-		boolean[] array = new boolean[this.size()];
-		
-		for(int index = 0; index < array.length; index++) {
-			
-			array[index] = this.getBoolean(index);
-		}
-		
-		return array;
-	}
-	
-	/**
-	 * Converts this {@linkplain JSONArray} to a XML string.
-	 * @param rootName name of the root element
-	 * @return this JSON array as a XML string
-	 * @since 1.1.0
-	 */
-	public String toXML(String rootName) {
-		
-		StringBuilder builder = new StringBuilder();
-		builder.append('<');
-		builder.append(rootName);
-		builder.append(" length=");
-		builder.append(this.size());
-		builder.append('>');
-		
-		for(Object element : this) {
-
-				   if(element instanceof JSONObject) {builder.append(((JSONObject)element).toXML("item"));
-			} else if(element instanceof JSONArray) {builder.append(((JSONArray)element).toXML("item"));
-			} else {
-					
-				builder.append("<item>");
-				
-				if(element != null) {
-					
-					builder.append(String.valueOf(element));
-				}
-				
-				builder.append("</item>");
-			}
-		}
-		
-		builder.append("</");
-		builder.append(rootName);
-		builder.append('>');
-		
-		return builder.toString();
-	}
-}
