+++ src/main/java/org/ralleytn/simple/json/JSONContentHandler.java
@@ -201,63 +201,109 @@
  *    See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package de.ralleytn.simple.json.internal;
+package org.ralleytn.simple.json;
+
+import java.io.IOException;
 
 /**
- * Represents a token.
+ * A simplified and stoppable SAX-like content handler for stream processing of JSON text. 
+ * @see org.xml.sax.ContentHandler
+ * @see org.ralleytn.simple.json.JSONParser#parse(java.io.Reader, JSONContentHandler, boolean)
  * @author FangYidong(fangyidong@yahoo.com.cn)
- * @author Ralph Niemitz/RalleYTN(ralph.niemitz@gmx.de)
- * @version 2.0.0
+ * @version 1.0.0
  * @since 1.0.0
  */
-public class Yytoken {
-	
-	/** @since 1.0.0 */ public static final int TYPE_VALUE = 0;
-	/** @since 1.0.0 */ public static final int TYPE_LEFT_BRACE = 1;
-	/** @since 1.0.0 */ public static final int TYPE_RIGHT_BRACE = 2;
-	/** @since 1.0.0 */ public static final int TYPE_LEFT_SQUARE = 3;
-	/** @since 1.0.0 */ public static final int TYPE_RIGHT_SQUARE = 4;
-	/** @since 1.0.0 */ public static final int TYPE_COMMA = 5;
-	/** @since 1.0.0 */ public static final int TYPE_COLON = 6;
-	/** @since 1.0.0 */ public static final int TYPE_EOF = -1;
-	
-	public int type;
-	public Object value;
-	
+public interface JSONContentHandler {
 	/**
-	 * @param type the token type
-	 * @param value the value of this token
+	 * Receives a notification when the JSON processing begins.
+	 * The {@linkplain JSONParser} will invoke this method only once.
+	 * @throws JSONParseException {@linkplain JSONParser} will stop and throw the same {@linkplain Exception} to the caller when receiving this {@linkplain Exception}.
+	 * @throws IOException {@linkplain JSONParser} will stop and throw the same {@linkplain Exception} to the caller when receiving this {@linkplain Exception}.
 	 * @since 1.0.0
 	 */
-	public Yytoken(int type, Object value) {
-		
-		this.type = type;
-		this.value = value;
-	}
+	public void startJSON() throws JSONParseException, IOException;
 	
-	// ==== 17.03.2018 | Ralph Niemitz/RalleYTN(ralph.niemitz@gmx.de)
-	// Since this is an internal class and no classes inside this library use this method, it is dead code.
-	// ====
+	/**
+	 * Receives a notification when the JSON processing ends.
+	 * @throws JSONParseException {@linkplain JSONParser} will stop and throw the same {@linkplain Exception} to the caller when receiving this {@linkplain Exception}.
+	 * @throws IOException {@linkplain JSONParser} will stop and throw the same {@linkplain Exception} to the caller when receiving this {@linkplain Exception}.
+	 * @since 1.0.0
+	 */
+	public void endJSON() throws JSONParseException, IOException;
 	
-	/*@Override
-	public String toString() {
-		
-		StringBuilder builder = new StringBuilder();
-		
-		// ==== 01.03.2018 | Ralph Niemitz/RalleYTN(ralph.niemitz@gmx.de)
-		// Replaced switch-case with if-else because if-else is faster
-		// ====
-		
-		       if(this.type == Yytoken.TYPE_VALUE) {builder.append("VALUE(").append(this.value).append(")");
-		} else if(this.type == Yytoken.TYPE_LEFT_BRACE) {builder.append("LEFT BRACE({)");
-		} else if(this.type == Yytoken.TYPE_RIGHT_BRACE) {builder.append("RIGHT BRACE(})");
-		} else if(this.type == Yytoken.TYPE_LEFT_SQUARE) {builder.append("LEFT SQUARE([)");
-		} else if(this.type == Yytoken.TYPE_RIGHT_SQUARE) {builder.append("RIGHT SQUARE(])");
-		} else if(this.type == Yytoken.TYPE_COMMA) {builder.append("COMMA(,)");
-		} else if(this.type == Yytoken.TYPE_COLON) {builder.append("COLON(:)");
-		} else if(this.type == Yytoken.TYPE_EOF) {builder.append("END OF FILE");
-		}
-
-		return builder.toString();
-	}*/
+	/**
+	 * Receives a notification when a JSON object begins.
+	 * @return {@code false} if the {@linkplain JSONContentHandler} wants to stop parsing after return
+	 * @throws JSONParseException {@linkplain JSONParser} will stop and throw the same {@linkplain Exception} to the caller when receiving this {@linkplain Exception}.
+     * @throws IOException {@linkplain JSONParser} will stop and throw the same {@linkplain Exception} to the caller when receiving this {@linkplain Exception}.
+     * @see #endJSON
+     * @since 1.0.0
+	 */
+	public boolean startObject() throws JSONParseException, IOException;
+	
+	/**
+	 * Receives a notification when a JSON object ends.
+	 * @return {@code false} if the {@linkplain JSONContentHandler} wants to stop parsing after return
+	 * @throws JSONParseException {@linkplain JSONParser} will stop and throw the same {@linkplain Exception} to the caller when receiving this {@linkplain Exception}.
+     * @throws IOException {@linkplain JSONParser} will stop and throw the same {@linkplain Exception} to the caller when receiving this {@linkplain Exception}.
+     * @see #startObject
+     * @since 1.0.0
+	 */
+	public boolean endObject() throws JSONParseException, IOException;
+	
+	/**
+	 * Receives a notification when a JSON object entry begins.
+	 * @param key name of the object entry
+	 * @return {@code false} if the {@linkplain JSONContentHandler} wants to stop parsing after return
+	 * @throws JSONParseException {@linkplain JSONParser} will stop and throw the same {@linkplain Exception} to the caller when receiving this {@linkplain Exception}.
+     * @throws IOException {@linkplain JSONParser} will stop and throw the same {@linkplain Exception} to the caller when receiving this {@linkplain Exception}.
+     * @see #endObjectEntry
+     * @since 1.0.0
+	 */
+	public boolean startObjectEntry(String key) throws JSONParseException, IOException;
+	
+	/**
+	 * Receives a notification when a JSON entry ends.
+	 * @return {@code false} if the {@linkplain JSONContentHandler} wants to stop parsing after return
+	 * @throws JSONParseException {@linkplain JSONParser} will stop and throw the same {@linkplain Exception} to the caller when receiving this {@linkplain Exception}.
+     * @throws IOException {@linkplain JSONParser} will stop and throw the same {@linkplain Exception} to the caller when receiving this {@linkplain Exception}.
+     * @see #startObjectEntry
+	 */
+	public boolean endObjectEntry() throws JSONParseException, IOException;
+	
+	/**
+	 * Receives a notification when a JSON array begins.
+	 * @return {@code false} if the {@linkplain JSONContentHandler} wants to stop parsing after return
+	 * @throws JSONParseException {@linkplain JSONParser} will stop and throw the same {@linkplain Exception} to the caller when receiving this {@linkplain Exception}.
+     * @throws IOException {@linkplain JSONParser} will stop and throw the same {@linkplain Exception} to the caller when receiving this {@linkplain Exception}.
+     * @see #endArray
+     * @since 1.0.0
+	 */
+	public boolean startArray() throws JSONParseException, IOException;
+	
+	/**
+	 * Receives a notification when a JSON array ends.
+	 * @return {@code false} if the {@linkplain JSONContentHandler} wants to stop parsing after return
+	 * @throws JSONParseException {@linkplain JSONParser} will stop and throw the same {@linkplain Exception} to the caller when receiving this {@linkplain Exception}.
+     * @throws IOException {@linkplain JSONParser} will stop and throw the same {@linkplain Exception} to the caller when receiving this {@linkplain Exception}.
+     * @see #startArray
+     * @since 1.0.0
+	 */
+	public boolean endArray() throws JSONParseException, IOException;
+	
+	/**
+	 * Receives a notification when a primitive JSON value is reached:
+	 * <ul>
+	 * <li>{@linkplain String}</li>
+	 * <li>{@linkplain Number}</li>
+	 * <li>{@linkplain Boolean}</li>
+	 * <li>{@code null}</li>
+	 * </ul>
+	 * @param value the primitive JSON value
+	 * @return {@code false} if the {@linkplain JSONContentHandler} wants to stop parsing after return
+	 * @throws JSONParseException {@linkplain JSONParser} will stop and throw the same {@linkplain Exception} to the caller when receiving this {@linkplain Exception}.
+	 * @throws IOException {@linkplain JSONParser} will stop and throw the same {@linkplain Exception} to the caller when receiving this {@linkplain Exception}.
+	 * @since 1.0.0
+	 */
+	public boolean primitive(Object value) throws JSONParseException, IOException;
 }
