+++ src/main/java/org/ralleytn/simple/json/internal/Yylex.java
@@ -201,14 +201,14 @@
  *    See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package de.ralleytn.simple.json.internal;
+package org.ralleytn.simple.json.internal;
 
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.io.Reader;
 
-import de.ralleytn.simple.json.JSONParseException;
+import org.ralleytn.simple.json.JSONParseException;
 
 /* The following code was generated by JFlex 1.4.2 */
